requirements.txt
asgiref==3.10.0
Django==5.2.7
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
